Fix Windows 64-bit compilation error.
BUG=
TEST=

Review URL: http://codereview.chromium.org/7307002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@8525 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -5759,7 +5759,7 @@ void FindAsciiStringIndices(Vector<const char> subject,
     pos = reinterpret_cast<const char*>(
         memchr(pos, pattern, subject_end - pos));
     if (pos == NULL) return;
-    indices->Add(pos - subject_start);
+    indices->Add(static_cast<int>(pos - subject_start));
     pos++;
     limit--;
   }
